+++ src/apibara/__init__.py
@@ -1,4 +1,5 @@
 __version__ = "0.5.0"
 
 from .indexer.runner import IndexerRunner, Info
-from .model import Event, EventFilter, NewBlock, NewEvents, Reorg, StarkNetEvent
+from .model import (Event, EventFilter, NewBlock, NewEvents, Reorg,
+                    StarkNetEvent)